Add included hook, so include TinySingleton works
In fact, it is not strictly necessary to use Module#append.
We need to intercept Object#clone and Object#dup (on the object level) and
Class#new on the class level. So long that our modules TinySingleton and
TinySingleton::ClassMethods are before Object and Class in the ancestor
chain respectively it should be alright.

Using Module#append is more of a failsafe in case some other modules that
play with clone/dup are included.